moving things under /aemedge
